confd: initial audit trail support
This commit introduces a new SECURITY() log macro that logs to facility
'audit' (sometimes called 'security') with severity set to 'alert'.  It
is then used to instrument the single most important things to log; all
changes to users and their privileges.

We also add logging to the core_startup_save() and add a new callback
for core_running_save() to log any changes made to startup-config and
running-config by a user.

Fixes #521

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>